Workaround a bug in QDirModel in tst_QDirModel::task196768_sorting
The test was reported failing by pulse on many platform, but noone of
us was able to reproduce it locally.  The problem was that pulse runs the
tests in a hidden dirrectory.  And this exposed a bug in the QDirModel.

We decided anyway that the bug was not worth fixing considered that
QDirModel is kind of deprecated in favor of QFileSystemModel.

(The bug is the one reported in task 255066)

Reviewed-by: mbm
Reviewed-by: Thierry