Fix "... may return wrong private dir if application is on Z:, not C:"
QT-805 QCoreApplication::applicationDirPath() may return wrong private
dir if application is on Z:, not C:

Remove the emulator ifdef, instead RFs::Drive is used to get the drive
attributes (to see if it is ROM / read only)
RFs::SystemDrive is used to get the system drive, which is C:
on S60 phones, but in theory configurable.
It is not needed to call Open() on RProcess to call functions on the
current process - this eliminates some code.
Elimated some unneeded code to check for Z: before creating prvate path
(now, the if statement would always have evaluated the same)

Task-number: QT-805
Reviewed-by: Miikka Heikkinen