Modify functions for native pixmap data conversion on Symbian.
We need a way to support converion to and from multiple native pixmap
types from multiple pixmap backends. Instead of adding more virtual
functions to QPixmapData, make the existing one more generic but pass
an opaque pointer and a type and do some internal casting.

Currently this function is Symbian only, but could easily be extended
to work on other platforms.

Reviewed-by: Aleksandar Babic
Reviewed-by: Jani Hautakangas