Fix regressions in qeventloop, qtimer, and qsocketnotifier autotests
Commit ed375675d4a4f6fd63edeb242e23c87b3de4be6f triggers a behavior in
Glib's mainloop implementation where some event sources are not
"serviced" every iteration of the mainloop context. This breaks an
invariant that many tests relied on, so we need to solve the problem.

The invariant is that a newly added timer that would normally fire on
the next pass of the event loop (liker a zero timer) SHOULD actually
fire. We do this by registering 2 timer event sources with Glib's
mainloop: one normal priority source and one idle priority source. The
idle priority source is the one that will send events most of the
time, with the normal priority one taking over only when
processEvents() is called manually.

Task-number: QT-877
Reviewed-by: jbache
Reviewed-by: thiago
Reviewed-by: denis