Made fast blur in GL 2 engine be radius independent.
This is useful when animating the blur radius, as you don't want to
suffer the hit of compiling / linking a new shader program for each
radius the first time the animation is played.

Also use the fast blur when the radius is 5 or below, as the quality
difference is insignificant.

Reviewed-by: Rhys Weatherley