Try to support Qt 4.4-style Phonon includes in Qt.
This introduces an undocumented "phonon_compat" subdir and also adds
it to the INCLUDEPATH when QT += phonon is specified. With this, these
styles of #includes should be supported:

 #include <Phonon>
 #include <Phonon/Phonon>
 #include <phonon>
 #include <phonon/phonon>
 #include <Phonon/MediaObject>
 #include <phonon/mediaobject.h>

Still need to check if the headers get installed during "make
install". I couldn't find where in our code that is done.

BT: yes
Reviewed-By: Trust Me