Avoid calling ensureContext() in setGeometry.
We can delay calling ensureContext() since it will happen when painting
begins anyway. There is a chance that we can get to this function when
a window is being hidden and if this is the first thing that happens in
the application, there is no need to call ensureContext() yet.

Reviewed-by: Rhys Weatherley