QGraphicsAnchorLayout: Enabling Simplex support for Expanding size policy
To support the expanding size policy we had to change the way the setup
process work. Previously we used to calculate three key-frames using
the simplex solver:

 - sizeAtMinimum: the value an anchor should be in when the layout is
                  at its minimum size. Calculated using simplex solver
	          to minimize the layout size.
 - sizeAtPreferred: the value an anchor should be in when the layout is
                    at its preferred size. Calculated using simplex solver
	            to minimize the deviation from the items preferred
	            sizes.
 - sizeAtMaximum: the value an anchor should be in when the layout is
	          at its maximum size. Calculated using simplex solver
                  to maximize the layout size.

That worked fine but didn't diferentiate standard items from the
"expanding" ones. In other words, all items would grow from their
sizeAtPreferred to their sizeAtMaximum at the same rate.

To support the idea of "expanding" items, ie. items that should grow
faster than the others, we added a fourth state, between preferred
and maximum.

Now we have the following interpolation order:

 sizeAtMinimum -> sizeAtPreferred -> sizeAtExpanding -> sizeAtMaximum.

The definition of the "expanding" state is that all "expanding" items
should have grown all the way to their "sizeAtMaximum" values whereas
non expanding items should have kept their preferred sizes. The only
exception is that non-expanding items are allowed to grow if that is
necessary to allow the "expanding" items to reach their sizeAtMaximum.

So, the visual result is that if the layout is resized from its
preferred size to its maximum size, the expanding items will grow
first and then, in a second phase, the other items will grow.

This commit adds QGALPrivate::solveExpanding() and calls it from
calculateGraphs().

Signed-off-by: Eduardo M. Fleury <eduardo.fleury@openbossa.org>
Reviewed-by: Artur Duque de Souza <artur.souza@openbossa.org>