Only account for a negative right bearing in QTextLayout
We only want to expand the width of the text, never shrink it. This is
so that we can account for text where the pixels of the text extend
beyond the edge of the last glyph. When we shrinked the width, code
that depended on the natural text width equalling the advance of the
text (such as WebKit) broke, and some text items would be positioned
wrongly. We now only take negative right bearings into account (right
bearing is left-bound), meaning that we only expand the text width,
never shrink it due to right bearing.

Reviewed-by: Simon Hausmann