Fix the area above the menubar to have the right background
When a gradient was set on the whole window then there was a line above
the menubar that repeated the gradient for each menu item.  Since there
is a workaround in place already for the background of the menu to get
painted correctly, we just extend the rect this covers to ensure the
extra line of pixels is included.

Reviewed-by: Jens Bache-Wiig