Push the "lazy doneCurrent" logic down from QtOpenVG into QEgl
QtOpenVG was doing a lot of housekeeping to avoid having to
switch EGL contexts if the same surface was used over and over.
This housekeeping really belongs in the QEgl layer so that
QtOpenGL could potentially use it as well.

This change also adds some overrides for makeCurrent(),
swapBuffers(), and destroySurface() that take an EGLSurface
directly.  This is the first step in separating EGL surface
management from context management.

Reviewed-by: Sarah Smith