Make QEglContext API a little more flexible
Allow higher layers in QtOpenGL and QtOpenVG to set the EGLConfig
and EGLContext manually if they have some other way to determine
what the values should be (e.g. constructing a special config for a
specific platform).  Also add a QEglProperties argument to
createContext() to allow fine-tuning of the context parameters.

Reviewed-by: Sarah Smith