Stop qdoc outputting meaningless paths in html comments.
qdoc was generating the full path to the source document in a comment at
the top of each generated html file.  For source packages the path is a
temporary directory that only exists during package generation.  The
path also usually contains the Unix username of the user who generated
the package, and we shouldn't leak usernames for security and privacy
reasons.

This commit changes the code to write just the filename of the source
document into the comment instead of the full path.

Reviewed-by: Lincoln Ramsay