Make syncqt order headers deterministically.
Syncqt was relying on the order that headers appear in the filesystem.
That order is effectively non-deterministic, meaning that the files
generated when running syncqt on different copies of the same source are
very hard to compare.

For example, when generating source packages for the same sha1 twice in
a row, you would expect to get identical source packages, but you don't
because each packaging run downlaods the source tree from git, thus
randomizing the order of files on the filesystem.

This commit forces syncqt to be deterministic by making it sort each
directory it reads before it processes the directory.

On a side note: syncqt should probably be using Perl's File::Find
function instead of recursing through directories itself.

Reviewed-by: Lincoln Ramsay