Change the way we handle KeyboardUIMode on Mac
On Mac OS X, when the keyboard UI mode specifies "text boxes and lists
only", the tab key should only focus lists and text edit.

The previous implementation was using the focus policy to exclude the
buttons. This does not respect the configuration.

The change fixes tst_QApplication::focusChanged() with the Keyboard
mode "text boxes and lists only".

Reviewed-by: Richard Moe Gustavsen