Fix app showing a blank screen after switching to it using task switcher
On S60, the widget backing store is dropped when app window is hidden
in order to save memory. This has a side effect that the widgets are not
redrawn when the backing store is recreated if they have static content.

Fix is to invalidate the backing buffer after recreating it, and repaint
the widget.

Reviewed-by: Jason Barron