Changed S60 version check to do runtime check also in emulator.
S60 version check for emulator was hardcoded, which caused problems
for users of our binary packages, which deploy same binaries to all
environments.

Changed the check to look for platform .sis files also in emulator and
use the hardcoded version only as a fallback option for those SDKs
that do not provide these files.

Task-number: QTBUG-4579
Reviewed-by: axis