Make sure signal handlers are installed on QWS
QWS already installs message handlers, and this would cause testlib
not to install its own handler, causing lots of extra test failures
if there is a crash. The proper way would be to make testlib invoke
previous handlers, but the quick and easy fix is just to forget about
the old ones.

The rationale is that the worst thing that can happen is that it will
cause a crash, and that was going to happen anyway.

Reviewed-by: Jeremy