Mac: build issue (autotest stalls build)
When building Qt static on Mac, the test included actually stops
the build process, prompting the user for 'Promteroo?'
The reason is that configure runs qmake on all the tests inside
the test/auto directory, including the one that is meant for
testing qmake itself. In other words, this test should not
be qmake'ed when running configure, only when running the
qmake auto test. However, the qmake auto test does not run
the prompt test anymore either. So the solution for now
is to just comment out the test project as well.

Reviewed-by: alexis