Fixed QMessageBox::updateSize test case for S60.
In S60 message boxes are always in portrait mode located on the bottom
of screen and they occypy the whole sreen width. In addition
messageboxes in S60 does not have their own buttons, but the buttons are
located in system softkey area. That's why adding standard buttons for
message box in S60 does not change the dialog size.

Reviewed-by: Sami Merila