Fix tst_QTextDocument::cursorPositionChangedOnSetText
The test was introduced to make sure that the cursor position changed signal
was not emitted excessively on setPlainText or on setHtml. The original
fix however still included one superfluous emission for a temporary
QTextCursor object. This was fixed by change 930ba91ec1e630, this change
adjusts the auto test accordingly.

Reviewed-by: Simon Hausmann