QScript: Fix strange bugs and crashes.
I was assuming that the default return value register was always set
to 0 for native calls. But this is not the case.  So we must ensure this.

Also be consistend in the way the stackframe grow and shrink. This expose
another bug in the way the call frame is created in JSC

Reviewed-by: Kent Hansen