Add a more proper shutdown sequence for Symbian applications.
Previously both EEikCmdExit and EAknSoftkeyExit were hard coded to call
exit() which promptly exited the event loop without the chance to do
any UI operations. With this patch, we handle the two cases
differently and allow people to essentially make the right soft key
a hide instead of an exit simply by reimplementing closeEvent() and
ignoring the event.

EAknSoftkeyExit is now a soft exit since this does not seem to be a
mandatory exit in S60. By sending a QCloseEvent to QApplication this
has the effect that all windows are closed (sent another QCloseEvent)
and if one of them chooses to ignore this event, the application and
UI are both kept alive. This is similar to how Qt behaves on other
platforms. If all windows accept the close event (default), then the
application will quit() as normal. This command is sent from the
right soft key.

EEikCmdExit is mapped similarly as before, but calls quit() instead of
exit() to be aligned with other platforms. This will cause the
aboutToQuit signal to be emitted which maps very well to Qt because UI
interaction is not permitted from this signal and wouldn't work in S60
anyway because in the case of EEikCmdExit, the UI has already been torn
down by the UI framework. This command is sent by the End key.

We could take this one step further and intercept the key event and
prevent it from going to the UI framework and tearing down the UI,
but this is a little too evil IMHO. If people really want to
intercept that event, they can do it by reimplementing the platform
specific event filter. This is equivalent to how it would be done in
S60 (reimplementing HandleWsEvent).

Reviewed-by: axis
Reviewed-by: mread