Restoring cases that have been lost.
Due to bad merge some of the cases were removed.
They have been restored now.

Some adaptation for Symbian platform was needed as well.
We must not forget that local socket implementation is
based on RSocket, and therefore similar to having TCP_SOCKET
as basckend. So, in soe places it we had to give some time
for events to propagate by introducing qTest::wait()
lines.