Fix bug with posting events from HandleCommandL on Symbian.
Increase the loop level counter here because this is a system event
entry point. Previously, posting a deferred delete event from this
function (or triggering an action that posted this event) would end up
with the event not being processed on time. The reason is that these
events are only processed when the event loop that posted the event
returns.  Since the loop counter was not previously incremented, this
"nested loop" was not detected so the event would only be processed
when the main event loop returned which is typically at the end of the
application.

Reviewed-by: axis