Fix auto-test failure for QFileSystemModel::sort
This test was failing on the farm because the rootPath for the model
was invalid. QDir::rootPath on Windows return C:\ but in the farm
the temp directory is in E:\ therefore the sort was only triggered on C:\
and this explain the failure.

I also make comparaisons a bit more robust.

Reviewed-by:TrustMe