Make QPF's implementation of alphaMapForGlyph() consistent.
The QPF implementation of alphaMapForGlyph() was returning
color values of RGBA = (a, a, a, 255) instead of (0, 0, 0, a),
which was inconsistent with all the other font engines.

This inconsistency caused some QPF-specific workarounds in the
OpenGL and OpenVG paint engines.  This change removes the workarounds
and makes QPF generate the right colors from the start.  Paint
engines that ignore the color table or which don't use the
alphaMapForGlyph() function (e.g. raster) are not affected.

Reviewed-by: Paul