Fix tst_QFontDialog::task256466_wrongStyle() autotest failure in Cocoa.
Some fonts contain only styles other than Normal (or Regualr). If we try
to retrive the font sizes for such fonts by passing an empty style
string, the QFontDatabase will return a null list. This was causing the
autotest to fail. This patch will make sure that a style is always
selected in the QFontDialog.

Reviewed-by: Olivier