Ftp example application crashes (due to keypad navigation)
Keypad navigation tries to calculate the minimal distance of next
widget to the direction of pressed navigation key. This calculation
in QWidgetPrivate::widgetInNavigationDirection dies not take into
account that some widgets might have focusProxy setup.

In the reported case, ignoring focus proxy means that QDialogButtonBox
gets the focus and it hands it over to first button in its tab order.
Unfortunately, this button is disabled 'Download' button. Now, when
Select key is pressed, button action is triggered causing a crash.

Solution is to skip widgets that have focus proxies
in widgetInNavigationDirection.

Task-number: QT-2177
Reviewed-by: Alessandro Portale