Making tst_QListWidget::closePersistentEditor work for Symbian.
The persistent editor was not closed in time, as events from
Symbian app start up had not had time to be flushed through.
The addition of a 1s QTest::qWait gives plenty of time for the
app to settle.

Probably the better fix would be to add piece of code to testlib
which would empty the event queue before executing each test case.
However this was seen as an risky solution.

Reviewed-by: mread