Removed dummy VideoOutput instance.
This is part of an attempt to get the draggablevideo test app working on Symbian.  Use of the dummy video output object causes a top-level window to be created, which was suspected of being the reason why video is not visible.  This is not the case - for some reason, when the VideoOutput window is activated, it is still marked as 'hidden' by the window server (CWsClientWindow::ResetHiddenFlag, triggered from CCoeControl::ActivateL).