Fix an issue that was uncovered by two tst_QAbstractItemView fails.
tst_QAbstractItemView::shiftArrowSelectionAfterScrolling() and
tst_QAbstractItemView::shiftSelectionAfterRubberbandSelection() are two
relatively new test cases that always failed when Qt was configured
with QT_KEYPAD_NAVIGATION. And that even though we set the navigation
mode to Qt::NavigationModeNone when running autotests for Symbian.

Fix: in QAbstractItemViewPrivate::extendedSelectionCommand(), we make
sure that even with keypad navigation enabled, the shift-arrow feature
is available.

Reviewed-by: Jani Hautakangas

	modified:   src/gui/itemviews/qabstractitemview.cpp