Reset timeout-checker before evaluating when processEventsInterval is used
When the timeout-checker was not reset, it could take a while
(e.g. 1-2 seconds) before the next timeout occurred, depending on
what the tick counter happened to be after the previous evaluate().
When a processEventsInterval of e.g. 100 milliseconds has been
specified, we want the timeout to happen much sooner, thus we need
to reset the checker. This will cause the first timeout to happen
quickly, and then at steady intervals (processEventsInterval ms)
after that.

The tst_QScriptEngine::processEventsWhileRunning() test was
sporadically failing due to this issue.

Reviewed-by: Olivier Goffart