Map redraw rectangle into co-ordinate system of backing store, for native child widgets
The implementation of QSymbianControl::Draw previously assumed that the co-ordinate system of the control always matched that of the backing store - in other words, that the only widget for which a QSymbianControl was created would be the top-level widget.  Now that create_sys has been modified to create a native window (and therefore a QSymbianControl) for child widgets on which winId() is called, this assumption is no longer valid.  This function has therefore been modified to map the redraw rectangle into the coordinate system of the backing store.