Fix a bug when clipsChildrenToShape is set back to false.
If you set the flag itemClipsChildrenToShape to true on a parent, an
optimization was made in 4.5.0 to not add children of this parent in
the index. But when you set the flag back to false all the sub-tree
of the parent should be re-added to the index otherwise the index will
never find all children. This code is not relevant for 4.6 since
the index part of QGraphicsView has been refactored and handle this case
with itemChange in QGraphicsSceneIndex.

Reviewed-by:andreas