Mac: update/small fix to 45f095b8970dc3c1b6f6e97fa2323654ba848288
It turns out that we need to let singleStep keep its value
in qtextedit so that pushing the up/down buttons on the slider
works as they should. Moreover, overriding the behaviour in
abstract slider also makes QGraphicsView work out of the box.
Also added in the test fix suggested by Olivier.