Make it possible to implement engine-specific drop shadow filters
It wasn't possible for the paint engine to override the drop
shadow filter because QPixmapDropShadowFilter::draw() was not
asking for an engine-specific filter object.

Also, change the OpenVG filter to use vgGaussianBlur() instead of
vgConvolve(), and draw the original image on top of the shadow.

Task-number: QTBUG-4583
Reviewed-by: trustme
(cherry picked from commit 3909d97e86d62fd94e149925b5f3111c8f391334)