Disable etched text on rich text labels
When rendering etched text on disabled rich text labels
we would previously draw distorted links.
This is basically because anything that
did not render with the text foreground role would
be drawn twice as a shadow. Since there is no way to properly
render this text etched we will rather disable it completely
when the label uses rich text.

Task-number: QTBUG-4543
Reviewed-by: Simon Hausmann