Make QEventTransition works with QGraphicsObject.
QStateMachine framework installs QObject event filters to catch events
in order to triggers the proper transition. But installing a QObject event
filter on a QGraphicsObject gives nothing because QGraphicsView events
filters works differently. In order to make this works we now post
events using QApplication::postEvent in addition to the QGraphicsView
events.

Reviewed-by:andreas