Fixed noisy looking textures in Basic Drawing example with GL2 engine.
When running the Basic Drawing example with the GL2 paint engine and
antialiasing enabled, textures looked noisy. When antialiasing was
enabled, the example translated the painter half a pixel to get sharp
lines. This caused the textures to be sampled at texel corners. Without
linear interpolation, sampling textures at texel corners is
unpredictable. The fix is to not translate the painter.

Reviewed-by: Gunnar