Fixes to qpixmapcache test cases, test case now adapts to cache_limit.
The default cache_limit in different platforms have different value.
For example in Symbian the default is currently 1024 KB where as for
desktop platforms the default is 10 MB.

The purpose of modified qpixmapcache test cases was to do operations
until cache_limit was reached. This was achieved by hard coded 40000
iterations. However this hard-coded value is fargile for cache limit
changes, and in addition it unnecessarily made the test exectuion to
take very long time on platforms which had smaller cache limit.

This patch changes the test so that number of expected items to fit
in cache is calculated and then 1000 extra items is tried to put
in cache to make sure limit is exceeded.

Reviewed-by: Alexis Menard
(cherry picked from commit bfe0e5c0d47780542b174dc96973920fba11c451)