Fixed a crash in menus on Symbian.
The reason for the crash was the following: When we make menu entries
in Qt, we assign each item an arbitrary command ID. This is because
Symbian usually puts the items in a resource file and refers to them
by ID, but we need to be dynamic. These command IDs are also assigned
to cascading menu items (sub menus). When we then get a callback in
RestoreMenuL with one of submenu IDs, we used to ask Symbian to
construct the menu items for them, but Symbian doesn't know about
them.

Fixed by avoiding call into S60 code if the ID belongs to Qt.

Also put a cap on the number of menu items. It's very unlikely that
anyone will reach it, but it's better to have an actual check.

Task:     QT-646
AutoTest: Manual testing went fine
RevBy:    mread