QScript: do not crash on PowerPC
There is no 'this' register in the global context.
The computation of the this register for the global context
gives the 'codeBlock' register in the frame header.

On Intel processor, a JSValue() is 0x0 when converted to a pointer,
but this is not the case on PowerPC  (it is 0xfffffff9) so it just
crash later when acessing the code block.

Solution: special condition for the global context when getting the
'this' object

Reviewed-by: Kent Hansen
(cherry picked from commit 37bd7a5711e57ea8c45ae75102ddee3ab905a0e5)