Fixed QAudioDeviceInfoInternal::deviceList(QAudio::Mode mode) filtering
If an audio device only supported Input or Output, it would not be added
to the list of devices. Only devices that returned IOID == NULL would be
added. Also, _m was not being used, and io was unneccessarily being cast
to a QString.

Merge-request: 1664
Reviewed-by: Kurt Korbatits <kurt.korbatits@nokia.com>
(cherry picked from commit ba11343826229e983cb1d38811c8363d355e2e01)