Fix the pad navigator demo.
QGraphicsWidget used to called setPosHelper where all the logic was.
But since the new flag itemSendsGeometryChanges some part of the code
inside setPosHelper move back to setPos. QGraphicsWidget was not updated
after this change. It doesn't matter as it is but for QGraphicsProxyWidget
which activate the flag itemSendsGeometryChanges it matters. ItemChange
was never called so the proxy was never really moved.

Task-number:QT-672
Reviewed-by:andreas
(cherry picked from commit 3da33626c056169f5fadf94f12997180cb3a08b4)