Add GNOME implementation for native filesystem icons
This adds some code to support native filesystem icons
on GNOME. It works by resolving gnome libs and gnome-vfs dynamically,
hence we are explicitly running it on GNOME only and not KDE.
Even if it would work there as well. We are planning on adding this
functionality to the platform plugin as well.

Task-number: QTBUG-2195
Reviewed-by: joao