Avoids timer ticks when there are only pause animations running
When there are only pause animations running, the timer will stop and
restart when the closest pause animation finishes. While there are only
pause animations running, there are no additional timer ticks, but
if there is at least one animation running that is not a group or a pause,
then the global animation timer will restore it's update interval.

Includes a new auto-test for the QPauseAnimation class.

Task-number: QT-941
Reviewed-by: thierry
Reviewed-by: janarve