Determine QPrinterInfo's supportedPaperSizes on demand.
The paper size determination requires a cupsGetPPD(), which can be a slow
HTTP download.
So doing it on demand makes QPrinterInfo::defaultPrinter() much faster,
because it doesn't actually need the paper sizes at all. Before this fix,
it was requesting the PPD file of every known CUPS printer, every time
it was called.
My battery of unittests for code that uses QPrinter went from 19 minutes
when the cups print server is switched off, to 13 seconds!

This fix also removes some code duplication, which is always nice.

Task-number: 232664, QTBUG-3033
Reviewed-by: Trond