Fixed deployment when using cetest.
Cetest (and other programs that upload dlls manually without using a
package) need to deploy some plugins for specific tests. If those
tests are deployed in a normal package however, the installation will
fail because the plugins are already included in the Qt installation.
Fixed that by putting the deployment inside a scope that cetest will
define.

RevBy:    Miikka Heikkinen