Avoid adding a debug library to the glue projects, when debug_and_release
The VCPROJ/SLN generator generates dependencies on the glue project,
so the glue must use the correct library for the dependencies to be
correct. The qtAddLibrary() would add the project 'default' to the
glue, which could end up adding a debug lib to the glue, so the
dependency checker wouldn't find the proper project.

We therefore force qtAddLibrary to only add d/_debug if we're not
using debug_and_release, or not in the glue part

Reviewed-by: Rohan McGovern