Optimized clipping in the GL 2 engine for winding fills.
When no stencil clip is set we reduce the number of compositing passes
for rendering winding fill paths from four to two. When stencil clip is
set, the number of compositing passes is reduced from five to four.

For clipping with a winding fill path, the number of compositing passes
are reduced from five to four when stencil clipping is already enabled.

A performance improvement of up to 85 % was measured in certain cases.

Reviewed-by: Trond