Fix: Lazy instantiation of a static QAction
We should not create a QAction instance at program startup,
that is before QApplication was initialized. One reason not
to do that is that internal fonts are prematurely initialized
without QApplication::qt_is_gui_used set to true, which leads
to wrong font Dpis in qt_defaultDpiX().

This issue was detected due to the failure in tst_QTextLayout,
cases: testDefaultTabs, testTabs, testMultilineTab,
testRightTab, testTabsInAlignedParag, testCenteredTab,
testDelimiterTab, testMultiTab and tabsForRtl.

Fix: create a Q_GLOBAL_STATIC_WITH_ARGS for that QAction
instance.

Reviewed-by: Liang QI